fix filepath of wav file update the install script

+++ install.sh
@@ -2,8 +2,14 @@
 
 echo "Resolving dependencies..."
 
-sudo apt update
-sudo apt install sox -y
+FILE=/usr/bin/sox
+if test -f "$FILE"; then
+    echo "SoX already installed, moving on..."
+else
+    sudo apt update
+    sudo apt install sox -y
+fi
+
 git clone https://github.com/markondej/fm_transmitter ./remove_me_later
 cd remove_me_later || echo "Error while cloning repository, please check your git install"
 
@@ -22,3 +28,4 @@ echo "BE SURE TO POPULATE .ENV FILE"
 echo "============================="
 
 cp .env.example .env
+mkdir uploads
